+++ cmd/smyrna/selection.c
@@ -1033,7 +1033,7 @@ int point_within_sphere_with_coords(float x0, float y0, float z0, float r,
 float distance_to_line(float ax,float ay,float bx,float by,float cx,float cy)
 {
 	//this function returns the distance between a line(a-b) segment and a point(c) in 2D plane
-	return sqrt(
+	return (float)sqrt(
 				pow(((by-ay)*(cx-ax)+(bx-ax)*(cy-ay)),2)
 				/
 				(  pow((bx-ax),2) + pow((by-ay),2))
